Split extractor module and refine frame-handling tests
Decompose extractor into a submodule layout by moving concrete
extractor implementations into src/extractor/extractors.rs and
converting extractor to directory module layout.

This keeps src/extractor/mod.rs below the 400-line threshold while
preserving the public extractor API via re-exports.

Update frame-handling tests to add the required module-level //! docs,
introduce an rstest harness fixture, and inject it into both tests to
remove duplicated setup.
